Added context menus (long presses)
Only added for shows activity at the moment.  Also the menu
is just the same as the options menu.

Also changed the menus to the res/menu's folder, which implied
also chaning the code to use R.menu.*, rather than R.layout.*, as it
is apperently a best practice.  (Accroding to developr.android.com)